FIX: deleted the routes folder and updated the server.js
